+++ CHANGELOG.md
@@ -13,5 +13,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Retrieval of locations, climbs and attempts from firebase
 - Views for locations and climbs
 - Location, climb, attempt, sections and grade set creation
-- Location editing and deletion
+- Location and climb editing
+- Location deletion
 - App icon